backend/requirements.txt
azure-identity>=1.15.0
azure-ai-projects>=1.0.0b1
azure-cosmos>=4.7.0
azure-search-documents>=11.4.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
aiohttp>=3.9.0
